Create SpanTestCase
